+++ src/components/DatatableV2/stories/ColumnActions.stories.tsx
@@ -7,9 +7,6 @@ import Template, { columns } from './Template';
 export default {
   title: 'components/DatatableV2/ColumnActions',
   component: Datatable,
-  parameters: {
-    screenshot: { skip: true },
-  },
 } as ComponentMeta<typeof Datatable>;
 
 export const ColumnActionsEnabled: Story = Template.bind({});