+++ src/components/DatatableV2/stories/ColumnOrdering.stories.tsx
@@ -62,6 +62,9 @@ InitialColumnOrdering.args = {
     ],
   },
 };
+InitialColumnOrdering.parameters = {
+  screenshot: { skip: false },
+};
 
 /**
  * For managed state all columns has to be provided as a default value