+++ src/components/DatatableV2/stories/HeaderTooltip.stories.tsx
@@ -1,5 +1,7 @@
 import type { ComponentMeta } from '@storybook/react';
 import type { Story } from './Template';
+import type { DatatableColumn, DatatableHeader } from '../Datatable.types';
+import type { DataSource } from '../mocks/data';
 
 import Datatable from '../Datatable';
 import Template, { columns } from './Template';
@@ -12,7 +14,13 @@ export default {
   },
 } as ComponentMeta<typeof Datatable>;
 
-const TooltipComponent = ({ header, column }) => (
+const TooltipComponent = ({
+  header,
+  column,
+}: {
+  header: DatatableHeader<DataSource>;
+  column: DatatableColumn<DataSource>;
+}) => (
   <div>
     {column.columnDef.header} - {header.id}
   </div>