+++ src/components/DatatableV2/stories/RowActions.stories.tsx
@@ -46,17 +46,26 @@ RowActionsEnabled.args = {
     },
   ],
 };
+RowActionsEnabled.parameters = {
+  screenshot: { skip: false },
+};
 
 export const DisabledRowActions: Story = Template.bind({});
 DisabledRowActions.args = {
   ...Template.args,
   enableRowActions: false,
 };
+DisabledRowActions.parameters = {
+  screenshot: { skip: false },
+};
 
 export const NoRowActions: Story = Template.bind({});
 NoRowActions.args = {
   ...Template.args,
   enableRowActions: true,
+};
+NoRowActions.parameters = {
+  screenshot: { skip: false },
 };
 
 export const SingleRowAction: Story = Template.bind({});
@@ -74,6 +83,9 @@ SingleRowAction.args = {
     },
   ],
 };
+SingleRowAction.parameters = {
+  screenshot: { skip: false },
+};
 
 export const ConditionallyDisabledRowActions: Story = Template.bind({});
 ConditionallyDisabledRowActions.args = {
@@ -87,7 +99,7 @@ ConditionallyDisabledRowActions.args = {
         ({ row, table }) =>
         (event) =>
           action('row action')({ row, table, event }),
-      isDisabled: ({ row }) => row.original.organization.grade === 'A',
+      isDisabled: ({ row }) => row.original.organization.grade === 'D',
     },
     {
       iconName: 'lock',
@@ -121,7 +133,10 @@ ConditionallyDisabledSingleRowAction.args = {
         ({ row, table }) =>
         (event) =>
           action('row action')({ row, table, event }),
-      isDisabled: ({ row }) => row.original.organization.grade === 'A',
+      isDisabled: ({ row }) => row.original.organization.grade === 'D',
     },
   ],
+};
+ConditionallyDisabledSingleRowAction.parameters = {
+  screenshot: { skip: false },
 };