+++ src/components/DatatableV2/table/ProgressBar.tsx
@@ -1,6 +1,5 @@
 import styled, { keyframes } from 'styled-components';
 import { memo } from 'react';
-import { includes } from 'ramda';
 
 import { getRadii } from '../../../utils';
 
@@ -17,15 +16,13 @@ const wave = keyframes`
     transform: translateX(100%);
   }
 `;
-const ProgressBar = styled.span.withConfig({
-  shouldForwardProp: (property) => !includes(property, ['isTop', 'isBottom']),
-})<{ isTop?: boolean; isBottom?: boolean }>`
+const ProgressBar = styled.span<{ $isTop?: boolean; $isBottom?: boolean }>`
   display: block;
   background: #ece8ff;
   overflow: hidden;
   position: absolute;
-  inset: ${({ isBottom }) => (isBottom ? 0 : undefined)} 0
-    ${({ isTop }) => (isTop ? 0 : undefined)} 0;
+  inset: ${({ $isBottom }) => ($isBottom ? 0 : 'auto')} 0
+    ${({ $isTop }) => ($isTop ? 0 : 'auto')} 0;
   z-index: 1;
   height: 0.25rem;
   width: 100%;