+++ src/components/ScrollView/ScrollView.tsx
@@ -35,7 +35,8 @@ const ScrollViewRoot = styled(Surface)`
         rgb(0 0 0 / 15%),
         rgb(0 0 0 / 0%)
       )
-      center bottom;
+      center bottom,
+    var(--sscds-background);
   background-repeat: no-repeat;
   background-size: 100% 40px, 100% 40px, 100% 14px, 100% 14px;
   background-attachment: local, local, scroll, scroll;
@@ -46,7 +47,7 @@ export const ScrollView = forwardRef<HTMLDivElement, ScrollViewProps>(
     {
       children,
       maxHeight = '100%',
-      radius = 'sm',
+      radius = 'none',
       hasBorder = false,
       style,
       ...props