+++ src/components/forms/TextArea/TextArea.test.tsx
@@ -1,4 +1,5 @@
 import { fireEvent, render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 
 import TextArea from './TextArea';
 
@@ -12,5 +13,14 @@ describe('TextArea', () => {
 
       expect(textarea).toHaveValue('0123456789');
     });
+  });
+
+  it('should update char counter', async () => {
+    render(<TextArea maxLength={5} />);
+
+    const textarea = screen.getByRole('textbox');
+    await userEvent.type(textarea, 'aaa');
+
+    expect(screen.getByText('2')).toBeInTheDocument();
   });
 });