+++ src/components/index.ts
@@ -46,3 +46,4 @@ export * from './layout';
 export * from './forms';
 export * from './typographyLegacy';
 export * from './ScrollView';
+export * from './Skeleton';